Get github pruner for elevate-cpanel working
Case RE-924: Get github pruner for elevate-cpanel working

The pruner here needed to have strings defined for both the VM name to be able
and Key we create unique to each VM to be able to parse them out of OpemStack.

Ideally we would really want to move the defines of both to the secrets side
so that in the future no further edits would ever have to be made to the pruner
as we would simply change that at the project level since it could be valuable
to other code outside the pruner.

Changelog: